Convert nc_nlen and cn_namelen to u_char.
NAME_MAX is 255, so u_char is sufficient.

+++ sys/kern/vfs_cache.c
@@ -80,7 +80,7 @@ namecache_compare(struct namecache *n1, struct namecache *n2)
 	if (n1->nc_nlen == n2->nc_nlen)
 		return (memcmp(n1->nc_name, n2->nc_name, n1->nc_nlen));
 	else
-		return (n1->nc_nlen - n2->nc_nlen);
+		return ((int)n1->nc_nlen - (int)n2->nc_nlen);
 }
 
 RB_GENERATE(namecache_rb_cache, namecache, n_rbcache, namecache_compare);
